Reparing ENV and reward
Problem with reset function, sometimes it left the main window opend.
We resolved this with while cylce (We wait until it appears and then
close it).

AgarEnv never calculated reward for done=True, quick fix was made

Lastly, we added model trained on 50000 (But with previous mistakes in
the code, The next run, will not have these mistakes and we expect
better preformece).

By all running test so far, we should set target update to 10000 and run
at least 1 000 000 runs.